+++ akshare/__init__.py
@@ -2361,9 +2361,10 @@ amac_manager_cancelled_info # 中国证券投资基金业协会-信息公示-诚
 1.9.43 fix: fix stock_a_congestion_lg interface
 1.9.44 fix: fix stock_a_high_low_statistics interface
 1.9.45 add: add stock_fhps_detail_ths interface
+1.9.46 fix: fix stock_a_gxl_lg interface
 """
 
-__version__ = "1.9.45"
+__version__ = "1.9.46"
 __author__ = "AKFamily"
 
 import sys